requirements.txt
pytest-playwright
pytest-bdd
allure-pytest
markdown